Reimplement the libevent-based event loop.
The previous implementation was based on wrong assumptions on how the
PECL libevent extension works. Please note that this class must be used
with PECL libevent v0.0.5 since they fixed a bug in which "event_set()"
was resetting the event base.

Due to a bug, most likely in the PECL extension, this class is not able
to use file streams as it considers them invalid streams. There is an
open issue on the PECL bug tracker for the libevent project:

  https://bugs.php.net/bug.php?id=59728

Everything works fine with socket streams and STDIN, STDOUT, STDERR, etc.