[EventLoop] Fix issues when removing streams with libevent.
When doing two consecutive updates on an event resource inside the same
tick, pecl-libevent seems to get stuck without properly freeing up the
event. It is still unclear if it is something related to pecl-libevent
or libevent itself, but our new approach seems to fix this problem.

For now we just delete the previous resource and create a new one one
on the associated stream using the new flags.

This commit should fix issue #9.